Reorganizing tests to use pytest
