+++ .github/workflows/run-tests.yml
@@ -19,13 +19,9 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install poetry>=1.3.2
+          pip install poetry
           poetry config virtualenvs.create false
           poetry install --with dev --no-root --no-interaction --no-ansi
-      - name: Lint with black
+      - name: Run linters
         run: |
-          black --check .
-      - name: Unit test with coverage
-        run: |
-          coverage run -m unittest -k tests.unit
-          coverage report
+          ./make.sh lint
