+++ .gitignore
@@ -136,3 +136,4 @@ docker-compose.override.yml
 config/*.json
 !config/transcript_cleanup.json
 scratch
+isochrone_cache.sqlite
