+++ docker/Dockerfile.fasterwhisper
@@ -3,7 +3,7 @@
 #
 # PLEASE DO NOT EDIT IT DIRECTLY.
 #
-FROM nvidia/cuda:12.1.1-cudnn8-runtime-ubuntu22.04
+FROM nvidia/cuda:12.6.0-cudnn8-runtime-ubuntu24.04
 
 RUN apt-get update && \
     apt-get -y upgrade && \
@@ -12,13 +12,16 @@ RUN apt-get update && \
     python3 \
     python3-pip \
     python3-dev \
+    pipx \
     git \
     ffmpeg \
     sox \
     && \
     rm -rf /var/lib/apt/lists/*
 
-RUN pip3 install poetry>=1.3.2
+RUN pipx install poetry
+
+ENV PATH="/root/.local/bin:${PATH}"
 
 # Install Faster Whisper
 COPY docker/install-faster-whisper.sh /usr/local/bin/install-whisper.sh
@@ -38,7 +41,6 @@ RUN poetry config virtualenvs.create false && \
 
 COPY app app
 COPY config config
-COPY bin bin
 COPY docker/docker-entrypoint.sh /usr/local/bin/
 
 ENV PYTHONPATH=/src
