+++ docker/Dockerfile.whispercpp
@@ -26,7 +26,7 @@ ARG WHISPER_MODEL=base.en
 ENV WHISPER_MODEL=${WHISPER_MODEL}
 RUN models/download-ggml-model.sh $WHISPER_MODEL
 
-FROM ubuntu:22.04
+FROM ubuntu:24.04
 
 RUN apt-get update && \
     apt-get -y upgrade && \
@@ -35,13 +35,16 @@ RUN apt-get update && \
     python3 \
     python3-pip \
     python3-dev \
+    pipx \
     git \
     ffmpeg \
     sox \
     && \
     rm -rf /var/lib/apt/lists/*
 
-RUN pip3 install poetry>=1.3.2
+RUN pipx install poetry
+
+ENV PATH="/root/.local/bin:${PATH}"
 
 # Install Whisper.cpp
 COPY --from=whispercpp /whisper.cpp/main /usr/local/bin/whisper-cpp
@@ -56,7 +59,6 @@ RUN poetry config virtualenvs.create false && \
 
 COPY app app
 COPY config config
-COPY bin bin
 COPY docker/docker-entrypoint.sh /usr/local/bin/
 
 ENV PYTHONPATH=/src
