+++ docker/docker-entrypoint.sh
@@ -6,7 +6,7 @@ if [ "$1" = 'api' ]; then
     /bin/sh -c "while true; do find /tmp -type f -mmin +10 -delete; sleep 60; done" &
     disown
 
-    exec uvicorn app.api:app --host 0.0.0.0 --log-level ${UVICORN_LOG_LEVEL:-info}
+    exec poetry run uvicorn app.api:app --host 0.0.0.0 --log-level ${UVICORN_LOG_LEVEL:-info}
 elif [ "$1" = 'worker' ]; then
     # Clean up any old temp files
     /bin/sh -c "while true; do find /tmp -type f -mmin +10 -delete; sleep 60; done" &
@@ -26,14 +26,14 @@ elif [ "$1" = 'worker' ]; then
 
     export LD_LIBRARY_PATH=`python3 -c 'import os; import nvidia.cublas.lib; import nvidia.cudnn.lib; print(os.path.dirname(nvidia.cublas.lib.__file__) + ":" + os.path.dirname(nvidia.cudnn.lib.__file__))'`
 
-    exec celery --app=app.worker.celery worker \
+    exec poetry run celery --app=app.worker.celery worker \
         -P ${CELERY_POOL:-prefork} \
         -c ${CELERY_CONCURRENCY:-1} \
         -l ${CELERY_LOGLEVEL:-info} \
         -n $CELERY_HOSTNAME \
         -Q ${CELERY_QUEUES:-transcribe}
 elif [ "$1" = 'flower' ]; then
-    exec celery --app=app.worker.celery flower --port=5555
+    exec poetry run celery --app=app.worker.celery flower --port=5555
 fi
 
 exec "$@"
