+++ examples/transcribe.sh
