+++ pyproject.toml
@@ -26,7 +26,7 @@ python-datauri = "^1.1.0"
 psycopg = {extras = ["binary", "pool"], version = "^3.1.12"}
 sqlalchemy = "^2.0.22"
 geopy = "^2.4.1"
-google-generativeai = "^0.4.0"
+google-generativeai = "^0"
 deepgram-sdk = "^3.2.2"
 shapely = "^2.0.4"
 requests-cache = "^1.2.0"
@@ -42,6 +42,7 @@ celery-stubs = "^0.1.3"
 types-requests = "^2.32.0.20240712"
 types-cachetools = "^5.5.0.20240820"
 types-pytz = "^2024.1.0.20240417"
+pytest-dotenv = "^0.5.2"
 
 [build-system]
 requires = ["poetry-core"]
