+++ tests/test_e2e.py
@@ -1,5 +1,4 @@
 import json
-import logging
 import os
 import unittest
 from time import sleep
@@ -8,9 +7,13 @@ import requests
 from dotenv import load_dotenv
 from meilisearch.errors import MeilisearchApiError
 
-from app import search
+from app.search import search
 
 load_dotenv()
+
+original_s3_public_url = os.getenv("S3_PUBLIC_URL")
+
+load_dotenv(".env.testing.local", override=True)
 
 
 class TestEndToEnd(unittest.TestCase):
@@ -25,23 +28,6 @@ class TestEndToEnd(unittest.TestCase):
             if err.code != "index_not_found":
                 raise err
         index = search.create_or_update_index(client, index_name)
-
-        while True:
-            try:
-                requests.get(
-                    url=f"{os.getenv('API_BASE_URL')}/config/notifications.json",
-                    headers={"Authorization": f"Bearer {os.getenv('API_KEY')}"},
-                    timeout=5,
-                ).raise_for_status()
-                requests.get(
-                    url=f"{os.getenv('S3_PUBLIC_URL')}/init-complete", timeout=5
-                ).raise_for_status()
-                logging.info("Connected to API successfully.")
-                break
-            except Exception as e:
-                logging.error(e)
-                logging.info("Waiting for API to come online...")
-                sleep(1)
 
     def transcribe(
         self,
@@ -116,7 +102,7 @@ class TestEndToEnd(unittest.TestCase):
             isinstance(json.loads(result["hits"][0]["raw_transcript"]), list)
         )
 
-        r = requests.get(result["hits"][0]["raw_audio_url"])
+        r = requests.get(result["hits"][0]["raw_audio_url"].replace(original_s3_public_url, os.getenv("S3_PUBLIC_URL")))
         self.assertEqual(200, r.status_code)
         self.assertEqual("audio/mpeg", r.headers.get("content-type"))
 
@@ -149,7 +135,7 @@ class TestEndToEnd(unittest.TestCase):
             isinstance(json.loads(result["hits"][0]["raw_transcript"]), list)
         )
 
-        r = requests.get(result["hits"][0]["raw_audio_url"])
+        r = requests.get(result["hits"][0]["raw_audio_url"].replace(original_s3_public_url, os.getenv("S3_PUBLIC_URL")))
         self.assertEqual(200, r.status_code)
         self.assertEqual("audio/mpeg", r.headers.get("content-type"))
 
@@ -181,7 +167,7 @@ class TestEndToEnd(unittest.TestCase):
             isinstance(json.loads(result["hits"][0]["raw_transcript"]), list)
         )
 
-        r = requests.get(result["hits"][0]["raw_audio_url"])
+        r = requests.get(result["hits"][0]["raw_audio_url"].replace(original_s3_public_url, os.getenv("S3_PUBLIC_URL")))
         self.assertEqual(200, r.status_code)
         self.assertEqual("audio/mpeg", r.headers.get("content-type"))
 
@@ -215,7 +201,7 @@ class TestEndToEnd(unittest.TestCase):
             isinstance(json.loads(result["hits"][0]["raw_transcript"]), list)
         )
 
-        r = requests.get(result["hits"][0]["raw_audio_url"])
+        r = requests.get(result["hits"][0]["raw_audio_url"].replace(original_s3_public_url, os.getenv("S3_PUBLIC_URL")))
         self.assertEqual(200, r.status_code)
         self.assertEqual("audio/mpeg", r.headers.get("content-type"))
